refactor: move cleanup screen to its Activity
Removes CleanupFragment an moves its behavior to CleanupActivity.
