+++ ui-cleanup/src/main/java/fr/nihilus/music/ui/cleanup/CleanupActivity.kt
@@ -16,20 +16,93 @@
 
 package fr.nihilus.music.ui.cleanup
 
+import android.Manifest
+import android.app.Activity
 import android.os.Bundle
+import androidx.activity.compose.setContent
+import androidx.activity.result.IntentSenderRequest
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.activity.viewModels
+import androidx.appcompat.app.AppCompatActivity
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import dagger.hilt.android.AndroidEntryPoint
-import fr.nihilus.music.core.ui.base.BaseActivity
+import fr.nihilus.music.core.compose.theme.OdeonTheme
+import fr.nihilus.music.core.ui.observe
+import fr.nihilus.music.media.tracks.DeleteTracksResult
 
 /**
- * A simple activity that only displays the content of [CleanupFragment].
- * This is required to avoid displaying the player controls on the cleanup screen
- * where controlling playback becomes secondary.
+ * Lists tracks that could be deleted to free up device space.
+ *
+ * This feature is exposed as an Activity to avoid sharing display of player controls, which becomes
+ * secondary in the context of this feature.
  */
 @AndroidEntryPoint
-class CleanupActivity : BaseActivity() {
+class CleanupActivity : AppCompatActivity() {
+    private val viewModel by viewModels<CleanupViewModel>()
+
+    private val requestPermission = registerForActivityResult(
+        ActivityResultContracts.RequestPermission()
+    ) { permissionGranted ->
+        if (permissionGranted) {
+            viewModel.deleteSelected()
+        }
+    }
+
+    private val deleteMediaPopup = registerForActivityResult(
+        ActivityResultContracts.StartIntentSenderForResult()
+    ) { result ->
+        if (result.resultCode == Activity.RESULT_OK) {
+            viewModel.clearSelection()
+        }
+    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_cleanup)
+
+        viewModel.state.observe(this) { state ->
+            if (state.result != null) {
+                when (state.result) {
+                    is DeleteTracksResult.Deleted -> {}
+                    is DeleteTracksResult.RequiresPermission -> {
+                        requestPermission.launch(Manifest.permission.WRITE_EXTERNAL_STORAGE)
+                    }
+
+                    is DeleteTracksResult.RequiresUserConsent -> {
+                        deleteMediaPopup.launch(
+                            IntentSenderRequest.Builder(state.result.intent).build()
+                        )
+                    }
+                }
+                viewModel.acknowledgeResult()
+            }
+        }
+
+        setContent {
+            val state by viewModel.state.collectAsState()
+
+            OdeonTheme {
+                var requiresDeleteConsent by remember { mutableStateOf(false) }
+                CleanupScreen(
+                    tracks = state.tracks,
+                    selectedCount = state.selectedCount,
+                    selectedFreedBytes = state.selectedFreedBytes,
+                    toggleTrack = { track -> viewModel.toggleSelection(track.id) },
+                    clearSelection = { viewModel.clearSelection() },
+                    deleteSelection = { requiresDeleteConsent = true },
+                )
+
+                if (requiresDeleteConsent) {
+                    ConfirmDeleteDialog(
+                        deletedTrackCount = state.selectedCount,
+                        accept = { viewModel.deleteSelected() },
+                        cancel = { requiresDeleteConsent = false }
+                    )
+                }
+            }
+        }
     }
 }
