--- ui-cleanup/src/main/res/layout/activity_cleanup.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?><!--
-  ~ Copyright 2020 Thibault Seisel
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-<androidx.fragment.app.FragmentContainerView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/fragment_host"
-    android:name="fr.nihilus.music.ui.cleanup.CleanupFragment"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="fr.nihilus.music.ui.cleanup.CleanupActivity"
-    tools:layout="@layout/fragment_cleanup" />
